requirements.txt
numpy>=1.24.0
flask==3.0.0
pandas==2.1.0
openpyxl==3.1.2
PyJWT==2.8.0
Werkzeug==2.0.1
python-dotenv==0.19.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
openai==1.12.0 